+++ lib/web_meta_seo.dart
@@ -21,13 +21,11 @@ class WebMetaSEO implements MetaSEO {
   /// This method should be run before any meta seo method to run the package correctly
   /// Implement the interface
   @override
-  seoMetaConfig() {
+  config() {
     /// Define the ScriptElement
     ScriptElement script = ScriptElement();
-
     /// Define the id of the ScriptElement
     script.id = 'metaSEOScripts';
-
     /// Define the javascript code of the ScriptElement
     script.innerHtml = """
   function seoNameJS(name, content) {
@@ -39,7 +37,6 @@ class WebMetaSEO implements MetaSEO {
     meta.setAttribute('content', content);
     document.getElementsByTagName('head')[0].appendChild(meta);
   }
-
   function seoPropertyJS(property, content) {
     if(document.querySelector("[property='"+property+"']") !== null) {
       document.querySelector("[property='"+property+"']").remove();
@@ -49,7 +46,6 @@ class WebMetaSEO implements MetaSEO {
     meta.setAttribute('content', content);
     document.getElementsByTagName('head')[0].appendChild(meta);
   }
-
   function seoAttributeJS(key, val) {
     if(document.querySelector("[name='"+key+"']") !== null) {
       document.querySelector("[name='"+key+"']").remove();
@@ -58,7 +54,16 @@ class WebMetaSEO implements MetaSEO {
     meta.setAttribute(key, val);
     document.getElementsByTagName('head')[0].appendChild(meta);
   }
-  
+  function seoOpenGraphJS(property, content) {
+    if(document.querySelector("[property='"+property+"']") !== null) {
+      document.querySelector("[property='"+property+"']").remove();
+    }
+    var meta = document.createElement('meta');
+    meta.setAttribute('property', property);
+    meta.setAttribute('content', content);
+    meta.setAttribute('data-rh', 'true');
+    document.getElementsByTagName('head')[0].appendChild(meta);
+  }
   function seoRobotsJS(name, content) {
     var meta = document.createElement('meta');
     meta.setAttribute('name', name);
@@ -66,19 +71,16 @@ class WebMetaSEO implements MetaSEO {
     document.getElementsByTagName('head')[0].appendChild(meta);
   }
     """;
-
     /// Make loop in html file body to check of any node with the same id
     for (int i = 0; i < document.body!.children.length; i++) {
       /// Check if the id of the package is exists in the html document
       if (document.body!.children[i].id == 'metaSEOScripts') {
         /// Remove any node with the same id of the javascript functions
         document.body!.children[i].remove();
-
         /// Then break the loop after deleting
         break;
       }
     }
-
     /// Add new or replace the javascript needed functions to the end
     /// of the body of the html document
     document.body!.insertAdjacentElement('beforeEnd', script);
@@ -91,7 +93,7 @@ class WebMetaSEO implements MetaSEO {
   /// Add web mata data of other meta named with content
   /// Implement the interface
   @override
-  seoMetaNameContent({required String name, required String content}) {
+  nameContent({required String name, required String content}) {
     /// Call the javascript function with needed attributes
     js.context.callMethod('seoNameJS', [name, content]);
   }
@@ -103,7 +105,7 @@ class WebMetaSEO implements MetaSEO {
   /// Add web mata data of other meta property with content
   /// Implement the interface
   @override
-  seoMetaPropertyContent({required String property, required String content}) {
+  propertyContent({required String property, required String content}) {
     /// Call the javascript function with needed attributes
     js.context.callMethod('seoPropertyJS', [property, content]);
   }
@@ -115,7 +117,7 @@ class WebMetaSEO implements MetaSEO {
   /// Add web mata data of other meta key with value
   /// Implement the interface
   @override
-  seoMetaKeyValue({required String key, required String value}) {
+  keyValue({required String key, required String value}) {
     /// Call the javascript function with needed attributes
     js.context.callMethod('seoAttributeJS', [key, value]);
   }
@@ -126,7 +128,7 @@ class WebMetaSEO implements MetaSEO {
   /// Example: String? author = 'Eng Mouaz M AlShahmeh';
   /// Implement the interface
   @override
-  seoAuthor({required String author}) {
+  author({required String author}) {
     /// Call the javascript function with needed attributes
     js.context.callMethod('seoNameJS', ['author', author]);
   }
@@ -137,7 +139,7 @@ class WebMetaSEO implements MetaSEO {
   /// Example: String? description = 'Simple SEO description mata data';
   /// Implement the interface
   @override
-  seoDescription({required String description}) {
+  description({required String description}) {
     /// Call the javascript function with needed attributes
     js.context.callMethod('seoNameJS', ['description', description]);
   }
@@ -148,7 +150,7 @@ class WebMetaSEO implements MetaSEO {
   /// Example: String? keywords = 'Flutter, Dart, Simple SEO';
   /// Implement the interface
   @override
-  seoKeywords({required String keywords}) {
+  keywords({required String keywords}) {
     /// Call the javascript function with needed attributes
     js.context.callMethod('seoNameJS', ['keywords', keywords]);
   }
@@ -160,7 +162,7 @@ class WebMetaSEO implements MetaSEO {
   /// Example: String? viewport = 'width=device-width, initial-scale=1';
   /// Implement the interface
   @override
-  seoViewport({required String viewport}) {
+  viewport({required String viewport}) {
     /// Call the javascript function with needed attributes
     js.context.callMethod('seoNameJS', ['viewport', viewport]);
   }
@@ -171,7 +173,7 @@ class WebMetaSEO implements MetaSEO {
   /// Example: String? httpEquiv = 'X-UA-Compatible';
   /// Implement the interface
   @override
-  seoHttpEquiv({required String httpEquiv}) {
+  httpEquiv({required String httpEquiv}) {
     /// Call the javascript function with needed attributes
     js.context.callMethod('seoAttributeJS', ['http-equiv', httpEquiv]);
   }
@@ -182,9 +184,19 @@ class WebMetaSEO implements MetaSEO {
   /// Example: String? charset = 'UTF-8';
   /// Implement the interface
   @override
-  seoCharset({required String charset}) {
+  charset({required String charset}) {
     /// Call the javascript function with needed attributes
     js.context.callMethod('seoAttributeJS', ['charset', charset]);
+  }
+
+  /// Definition of [fb:app_id] meta tag attribute
+  /// https://developers.facebook.com/docs/sharing/webmasters/
+  /// Add web mata data of [facebookAppID] attribute
+  /// Implement the interface
+  @override
+  facebookAppID({required String facebookAppID}) {
+    /// Call the javascript function with needed attributes
+    js.context.callMethod('seoPropertyJS', ['fb:app_id', facebookAppID]);
   }
 
   /// Definition of [og:title] meta tag attribute
@@ -193,9 +205,9 @@ class WebMetaSEO implements MetaSEO {
   /// Add web mata data of [ogTitle] attribute
   /// Implement the interface
   @override
-  seoOGTitle({required String ogTitle}) {
+  ogTitle({required String ogTitle}) {
     /// Call the javascript function with needed attributes
-    js.context.callMethod('seoPropertyJS', ['og:title', ogTitle]);
+    js.context.callMethod('seoOpenGraphJS', ['og:title', ogTitle]);
   }
 
   /// Definition of [og:description] meta tag attribute
@@ -204,9 +216,9 @@ class WebMetaSEO implements MetaSEO {
   /// Add web mata data of [ogDescription] attribute
   /// Implement the interface
   @override
-  seoOGDescription({required String ogDescription}) {
+  ogDescription({required String ogDescription}) {
     /// Call the javascript function with needed attributes
-    js.context.callMethod('seoPropertyJS', ['og:description', ogDescription]);
+    js.context.callMethod('seoOpenGraphJS', ['og:description', ogDescription]);
   }
 
   /// Definition of [og:image] meta tag attribute
@@ -215,9 +227,9 @@ class WebMetaSEO implements MetaSEO {
   /// Add web mata data of [ogImage] attribute
   /// Implement the interface
   @override
-  seoOGImage({required String ogImage}) {
+  ogImage({required String ogImage}) {
     /// Call the javascript function with needed attributes
-    js.context.callMethod('seoPropertyJS', ['og:image', ogImage]);
+    js.context.callMethod('seoOpenGraphJS', ['og:image', ogImage]);
   }
 
   /// Definition of [twitter:card] meta tag attribute
@@ -225,43 +237,32 @@ class WebMetaSEO implements MetaSEO {
   /// Add web mata data of [twitterCard] attribute
   /// Implement the interface
   @override
-  seoTwitterCard({required TwitterCard twitterCard}) {
+  twitterCard({required TwitterCard twitterCard}) {
     /// Make switch loop according to twitterCard state
     switch (twitterCard) {
       /// If the case is summary then run the following
       case TwitterCard.summary:
-
         /// Call the javascript function with summary attribute
         js.context.callMethod('seoNameJS', ['twitter:card', 'summary']);
-
         /// Break the switch loop if done
         break;
-
       /// If the case is summaryLargeImage then run the following
       case TwitterCard.summaryLargeImage:
-
         /// Call the javascript function with summary_large_image attribute
         js.context
             .callMethod('seoNameJS', ['twitter:card', 'summary_large_image']);
-
         /// Break the switch loop if done
         break;
-
       /// If the case is app then run the following
       case TwitterCard.app:
-
         /// Call the javascript function with app attribute
         js.context.callMethod('seoNameJS', ['twitter:card', 'app']);
-
         /// Break the switch loop if done
         break;
-
       /// If the case is player then run the following
       case TwitterCard.player:
-
         /// Call the javascript function with player attribute
         js.context.callMethod('seoNameJS', ['twitter:card', 'player']);
-
         /// Break the switch loop if done
         break;
     }
@@ -272,7 +273,7 @@ class WebMetaSEO implements MetaSEO {
   /// Add web mata data of [twitterTitle] attribute
   /// Implement the interface
   @override
-  seoTwitterTitle({required String twitterTitle}) {
+  twitterTitle({required String twitterTitle}) {
     /// Call the javascript function with needed attributes
     js.context.callMethod('seoNameJS', ['twitter:title', twitterTitle]);
   }
@@ -282,7 +283,7 @@ class WebMetaSEO implements MetaSEO {
   /// Add web mata data of [twitterDescription] attribute
   /// Implement the interface
   @override
-  seoTwitterDescription({required String twitterDescription}) {
+  twitterDescription({required String twitterDescription}) {
     /// Call the javascript function with needed attributes
     js.context
         .callMethod('seoNameJS', ['twitter:description', twitterDescription]);
@@ -293,7 +294,7 @@ class WebMetaSEO implements MetaSEO {
   /// Add web mata data of [twitterImage] attribute
   /// Implement the interface
   @override
-  seoTwitterImage({required String twitterImage}) {
+  twitterImage({required String twitterImage}) {
     /// Call the javascript function with needed attributes
     js.context.callMethod('seoNameJS', ['twitter:image', twitterImage]);
   }
@@ -305,52 +306,38 @@ class WebMetaSEO implements MetaSEO {
   /// Add web mata data of [content] attribute
   /// Implement the interface
   @override
-  seoRobots({required RobotsName robotsName, required String content}) {
+  robots({required RobotsName robotsName, required String content}) {
     /// Make switch loop according to robotsName state
     switch (robotsName) {
       /// If the case is google then run the following
       case RobotsName.google:
-
         /// Call the javascript function with google attribute
         js.context.callMethod('seoRobotsJS', ['google', content]);
-
         /// Break the switch loop if done
         break;
-
       /// If the case is googleBot then run the following
       case RobotsName.googleBot:
-
         /// Call the javascript function with googleBot attribute
         js.context.callMethod('seoRobotsJS', ['googlebot', content]);
-
         /// Break the switch loop if done
         break;
-
       /// If the case is googleBotNews then run the following
       case RobotsName.googleBotNews:
-
         /// Call the javascript function with googleBotNews attribute
         js.context.callMethod('seoRobotsJS', ['googlebot-news', content]);
-
         /// Break the switch loop if done
         break;
-
       /// If the case is googleSiteVerification then run the following
       case RobotsName.googleSiteVerification:
-
         /// Call the javascript function with googleSiteVerification attribute
         js.context
             .callMethod('seoNameJS', ['google-site-verification', content]);
-
         /// Break the switch loop if done
         break;
-
       /// If the case is robots then run the following
       case RobotsName.robots:
-
         /// Call the javascript function with robots attribute
         js.context.callMethod('seoRobotsJS', ['robots', content]);
-
         /// Break the switch loop if done
         break;
     }